final feedforward calculation

+++ Charged Up 2023/src/main/java/frc/robot/FireBirdsUtils.java
@@ -11,7 +11,7 @@ public class FireBirdsUtils{
         return (int)((angleInRads) / (2*Math.PI) * encoderCPR);
     }
   
-    public double CTRESensorUnitsToRads (int angleInSensorUnits, int encoderCPR) {
+    public double CTRESensorUnitsToRads (double angleInSensorUnits, int encoderCPR) {
        
       return (double) ((angleInSensorUnits/encoderCPR)*2*Math.PI);
